Massive refactoring of LoadGenerator components to follow better class structure
