+++ src/main/resources/jenkins/plugin/randomjobbuilder/TrivialLoadGenerator/config.jelly
@@ -1,12 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
-    <!-- FIXME make a form that submits posts to doBegin/doEnd request handle to call generator start/stop -->
-    <!--<f:form action="begin" name="config" method="post">
-        <f:submit name="trigger" value="toggle"></f:submit>
-    </f:form>-->
-
-    <!-- For some reason we can't show ${it.active} etc etc -->
 
     <f:entry field="generatorId" title="${%Generator ID (unique and autoassigned)}"
              description="Unique ID for the generator, autogenerated and can't be modified once configured">
